# Backend/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LangPlug API",
    "description": "German Language Learning Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/videos/{video_id}/vocabulary": {
      "get": {
        "tags": [
          "videos"
        ],
        "summary": "Get Video Vocabulary",
        "description": "Get vocabulary words extracted from a video",
        "operationId": "get_video_vocabulary_api_videos__video_id__vocabulary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VocabularyWord"
                  },
                  "title": "Response Get Video Vocabulary Api Videos  Video Id  Vocabulary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vocabulary/stats": {
      "get": {
        "tags": [
          "vocabulary"
        ],
        "summary": "Get Vocabulary Stats Endpoint",
        "description": "Get vocabulary statistics for the current user",
        "operationId": "get_vocabulary_stats_endpoint_api_vocabulary_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/vocabulary/blocking-words": {
      "get": {
        "tags": [
          "vocabulary"
        ],
        "summary": "Get Blocking Words",
        "description": "Get top blocking words for a video segment",
        "operationId": "get_blocking_words_api_vocabulary_blocking_words_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "video_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Path"
            }
          },
          {
            "name": "segment_start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Segment Start"
            }
          },
          {
            "name": "segment_duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 300,
              "title": "Segment Duration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vocabulary/mark-known": {
      "post": {
        "tags": [
          "vocabulary"
        ],
        "summary": "Mark Word As Known",
        "description": "Mark a word as known or unknown",
        "operationId": "mark_word_as_known_api_vocabulary_mark_known_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkKnownRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/vocabulary/preload": {
      "post": {
        "tags": [
          "vocabulary"
        ],
        "summary": "Preload Vocabulary",
        "description": "Preload vocabulary data from text files into database (Admin only)",
        "operationId": "preload_vocabulary_api_vocabulary_preload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/vocabulary/library/stats": {
      "get": {
        "tags": [
          "vocabulary"
        ],
        "summary": "Get Vocabulary Stats",
        "description": "Get vocabulary statistics for all levels",
        "operationId": "get_vocabulary_stats_api_vocabulary_library_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/vocabulary/library/{level}": {
      "get": {
        "tags": [
          "vocabulary"
        ],
        "summary": "Get Vocabulary Level",
        "description": "Get all vocabulary words for a specific level with user's known status",
        "operationId": "get_vocabulary_level_api_vocabulary_library__level__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "level",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vocabulary/library/bulk-mark": {
      "post": {
        "tags": [
          "vocabulary"
        ],
        "summary": "Bulk Mark Level Known",
        "description": "Mark all words in a level as known or unknown",
        "operationId": "bulk_mark_level_known_api_vocabulary_library_bulk_mark_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkMarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkMarkRequest": {
        "properties": {
          "level": {
            "type": "string",
            "pattern": "^(A1|A2|B1|B2|C1|C2)$",
            "title": "Level",
            "description": "CEFR difficulty level to mark (A1, A2, B1, B2, C1, C2)"
          },
          "known": {
            "type": "boolean",
            "title": "Known",
            "description": "Whether to mark all words in the level as known (true) or unknown (false)"
          }
        },
        "type": "object",
        "required": [
          "level",
          "known"
        ],
        "title": "BulkMarkRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarkKnownRequest": {
        "properties": {
          "word": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Word",
            "description": "The word to mark as known/unknown"
          },
          "known": {
            "type": "boolean",
            "title": "Known",
            "description": "Whether to mark the word as known (true) or unknown (false)"
          }
        },
        "type": "object",
        "required": [
          "word",
          "known"
        ],
        "title": "MarkKnownRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VocabularyStats": {
        "properties": {
          "levels": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Levels",
            "description": "Statistics by CEFR level with total and known counts"
          },
          "total_words": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Words",
            "description": "Total number of vocabulary words across all levels"
          },
          "total_known": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Known",
            "description": "Total number of known words across all levels"
          }
        },
        "type": "object",
        "required": [
          "levels",
          "total_words",
          "total_known"
        ],
        "title": "VocabularyStats"
      },
      "VocabularyWord": {
        "properties": {
          "word": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Word",
            "description": "The vocabulary word"
          },
          "definition": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition",
            "description": "Definition of the word"
          },
          "difficulty_level": {
            "type": "string",
            "pattern": "^(A1|A2|B1|B2|C1|C2)$",
            "title": "Difficulty Level",
            "description": "CEFR difficulty level (A1, A2, B1, B2, C1, C2)"
          },
          "known": {
            "type": "boolean",
            "title": "Known",
            "description": "Whether the user knows this word",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "word",
          "difficulty_level"
        ],
        "title": "VocabularyWord"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}